Add command to search for articles by name
