+++ src/main.rs
@@ -11,8 +11,9 @@ use itertools::Itertools;
 use crate::{
     formats::{html::convert_page_to_html, markdown::convert_page_to_markdown, PageFormat},
     languages::{fetch_all_langs, format_lang_table},
+    open_search::{format_open_search_table, open_search_to_page_url_tupel},
     utils::{create_cache_page_path, page_cache_exists},
-    wiki_api::fetch_page,
+    wiki_api::{fetch_open_search, fetch_page},
 };
 
 mod categories;
@@ -20,6 +21,7 @@ mod cli;
 mod error;
 mod formats;
 mod languages;
+mod open_search;
 mod utils;
 mod wiki_api;
 
@@ -79,6 +81,17 @@ async fn main() -> Result<(), WikiError> {
             if !no_cache_write {
                 fs::write(&page_cache_path, out.as_bytes())?;
             }
+
+            println!("{out}");
+        }
+        Commands::Search {
+            search,
+            limit,
+            lang,
+        } => {
+            let search_rs = fetch_open_search(&search, &lang, limit).await?;
+            let name_url_pairs = open_search_to_page_url_tupel(&search_rs)?;
+            let out = format_open_search_table(&name_url_pairs);
 
             println!("{out}");
         }
